Update followJsonPath logic
The logic for followJsonPath has been updated. Previously if the first token of the JSON path matched the table name then it would be omitted. This lead to issues when evaluating conditions against fields of a table after it had been joined.

The logic has now been changed so that upon recieving 'undefined' when attempting to omit the first token, rather than returning undefined it will instead proceed to attempt a normal resolution of the JSON path (i.e. without omitting the first token).